[feature/#65] data bind in home page and fixed course API.

+++ backend/src/services/UserService.ts
@@ -122,7 +122,7 @@ export const updateUserService = async (
       throw error;
     }
     let userProfile: string = req.body.userProfile;
-    if (req.files) {
+    if (req.files && req.files.userProfile?.length > 0) {
       userProfile = req.files.userProfile[0].path.replaceAll("\\", "/");
       if (user.userProfile && user.userProfile != userProfile) {
         deleteFile(user.userProfile);
